pyproject.toml
[project]
name = "bainite-martensite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama==0.4.6",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "filelock==3.14.0",
    "fonttools==4.53.0",
    "fsspec==2024.5.0",
    "h5py==3.11.0",
    "imageio==2.37.2",
    "intel-openmp==2021.4.0",
    "jinja2==3.1.4",
    "joblib==1.4.2",
    "kiwisolver==1.4.5",
    "lazy-loader==0.4",
    "markupsafe==2.1.5",
    "matplotlib==3.9.0",
    "mkl==2021.4.0",
    "mpmath==1.3.0",
    "networkx==3.3",
    "nir==1.0.4",
    "nirtorch==1.0",
    "numpy==1.26.4",
    "opencv-python>=4.11.0.86",
    "owlready2==0.49",
    "packaging==24.0",
    "pandas==2.2.2",
    "pillow==10.3.0",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.1",
    "scikit-image==0.26.0",
    "scikit-learn==1.5.0",
    "scipy==1.13.1",
    "shapely==2.1.2",
    "six==1.16.0",
    "snntorch==0.9.1",
    "sympy==1.12.1",
    "tbb==2021.12.0",
    "threadpoolctl==3.5.0",
    "tifffile==2026.1.28",
    "torchvision==0.18.0",
    "tqdm==4.66.4",
    "typing-extensions==4.12.1",
    "tzdata==2024.1",
]
